agentic_ai_etl_sttm_local_fullscd/requirements.txt
streamlit==1.35.0
pandas>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
PyYAML>=6.0.1
openai>=0.28.0
